[refactor] move EuiProvider decorator + globalTypes to separate file
- cleaner, contain concerns in a single file
